Add files via upload
The first time I did this, I didn't understand why I couldn't see the numbers, not realizing that I just had to open the console in my browser. Reminder, guys: `alert` will literally bring up a pop-up to the user, showing the number increasing each time in this example. The negative thing is that the user has to close the pop-up for it to refresh. Whereas, with `console.log`, the code just executes, and you will find the count immediately but will only show it on the console... I'm sure I'll find some more interesting ways to publish the results for the user.